Memoize `Table` rows to improve performance
