+++ packages/ui-components/package.json
@@ -79,6 +79,7 @@
     "date-fns": "2.9.0",
     "downshift": "^7.0.0",
     "lodash": "4.17.21",
+    "memoize-one": "^6.0.0",
     "nanoid": "^3.3.0",
     "polished": "3.6.4",
     "react": "^18.0",
